+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## v0.2.0 (unreleased)
 
+ * Add `dbn-go-file parquet` tool for processing DBN files with commands:
+   * `metadata`
+   * `json`
+   * `spit`
+   * `parquet`
  * Fix `Mbp0Msg` structure and add tests.
- * Add `dbn-go-file` tool for processing DBN files.
  * Add `MakeCompressedReader` and `MakeCompressedWriter` zstd helpers
 
 ## v0.1.0 (2025-01-13)